Update BCSabreImpl.php - Added new EXTERNAL functionality
Modified the BCSabreImpl::queryRangeTR() method to check for the flag "SPLIT:<num mins>" to be located in any of the "free" available event descriptions.  If a SPLIT:## token is found, the time slot will be split up into ## minute sessions.

This allows one to use one event to become available for a range of times.

Thanks much for this awesome project, and I hope this add-on serves others as well!